support for accepting and modifying shares, now share editing is more or less complete

+++ pCloud/binapi.h
@@ -63,7 +63,7 @@ typedef struct {
 #define P_STR(name, val) {PARAM_STR, (uint16_t)strlen(name), strlen(val), (name), {(uint64_t)(val)}}
 #define P_LSTR(name, val, len) {PARAM_STR, (uint16_t)strlen(name), (uint32_t)(len), (name), {(uint64_t)(val)}}
 #define P_NUM(name, val) {PARAM_NUM, (uint16_t)strlen(name), 0, (name), {(val)}}
-#define P_BOOL(name, val) {PARAM_BOOL, (uint16_t)strlen(name), 0, (name), {(val)?1:0}}
+#define P_BOOL(name, val) {PARAM_BOOL, (uint16_t)strlen(name), 0, (name), {(uint64_t)((val)?1:0)}}
 
 #define send_command(sock, cmd, ...) \
   ({\
